+++ src/api/jobposting/index.ts
@@ -126,29 +126,6 @@ export const getJobById = async (id: number): Promise<ApiResponse> => {
   return data
 }
 
-export const getApplicantsList = async (id: number): Promise<ApiResponse> => {
-  let data: ApiResponse = {
-    success: false,
-    code: 0,
-    message: '',
-    results: undefined
-  }
-  const url = `/api/job-postings/${id}/applicants`
-
-  await api
-    .get(url)
-    .then((res) => {
-      console.log(res)
-      data = res.data
-    })
-    .catch((error) => {
-      console.log(error)
-      data = error.response.data as ApiResponse
-    })
-
-  return data
-}
-
 export const getJobsPublic = async (): Promise<ApiResponse<JobPostingPublicResponse>> => {
   let data: ApiResponse<JobPostingPublicResponse> = {
     success: false,
