+++ src/api/resume/index.ts
@@ -7,6 +7,29 @@ import type {
   CoverLetterDescriptionResponse,
 } from '@/types/resume/ResumeTypes'
 
+export const getApplicantsList = async (id: number): Promise<ApiResponse> => {
+  let data: ApiResponse = {
+    success: false,
+    code: 0,
+    message: '',
+    results: undefined
+  }
+  const url = `/api/jobposts/${id}/applies`
+
+  await api
+    .get(url)
+    .then((res) => {
+      console.log(res)
+      data = res.data
+    })
+    .catch((error) => {
+      console.log(error)
+      data = error.response.data as ApiResponse
+    })
+
+  return data
+}
+
 export const createResume = async (
   resumeData: ResumeCreateRequest,
   file?: File,
